added mock generator file
